Include user32 and advapi32 in libuv extension
The libuv extension ("filesystem") wasn't including user32 or advapi32
in the linked libraries on windows because they were "already included".
However those inclusions have been moved to the event extension instead
of the core, so if you build without the event extension they will
be missing.

This adds them so it will link if you are building the filesystem
extension but not the event extension.