Use "INODE" of actions for the "exemplar"
The exemplar is a context that serves the purpose of a historical
"paramlist".  The frame's keylist names the function's parameters,
while the value slots do double duty: if a slot is hidden then it is
the specialized value for that parameter, but if it is visible then
it is the acceptable types and parameter class (quoted, refinement, etc)
for how that parameter should be fulfilled.

Due to being cramped for space in the finite node struture, the exemplar
was previously stored as part of the "archetype" action.  However,
this meant that if the archetype was overwritten in something like a
HIJACK operation, the original exemplar would be lost.  This meant if
the frame of the hijacking function was incompatible there would be
no way of knowing how to transform the incompatible frame into a
compatible one, as there was no remaining knowledge of the form to
transform to.

The INFO slot is not used in action definitions, so a new facility for
taking advantage of this (the INODE) is leveraged as the new storage
for the exemplar.  This simplifies the storage of the PARTIALS
which previously had to be put into a linked list with the exemplar.

Overall this should represent a simplification and a speedup for the
access of parameter lists from action nodes.