Fix resize handling in on_read()
This fixes details that came out in stress testing of the on_read()
and on_read_alloc() functions.  One is that it wasn't passing the
right amount to the Extend_Series() function, because due to the name
it was passing the amount to extend by...not the desired free capacity.
In addition to correcting that, this makes the name more clear.

Also fixes a bug pertaining to receiving an error in on_read() before
on_read_alloc() ever gets called.