Eliminate concept of AWAKE functions on ports
The current thinking is to completely remove R3-Alpha's attempts at
doing an asynchronous event model, and move to one that allows people
to express more of their code in a synchronous style, even if it
can be run by a scheduler in parallel.  This means drawing inspiration
from Go's "goroutines" or ASYNC/AWAIT.

That would require the stackless branch to be merged, which is not
ready yet.  But in the short term, committing to the design goal means
reframing the existing HTTP/HTTPD/TLS code to be synchronous.

As a first step, this wipes out all the "AWAKE" routines...and makes
OPEN/CONNECT/READ/WRITE/CLOSE synchronous.