Fix FAIL and QUIT exit status, add CI Tests
Due to changes related to merging DO and IMPORT for the
module system, scripts which were failing or running QUIT with a nonzero
exit code were still giving zero at exit.  This was meaning there was
no alert to the GitHub Action regarding failure.

This tries to workaround the problem by adding to the clearly unfinished
feature of QUIT handling for the common code path for import.  But also
adds in tests designed to fail for the CI to make sure they give the
right failing exit codes.