Make ODBC's HENV a global variable
Before you can make ODBC calls you need to create an HENV.  A HENV
is associated with an ODBC version (e.g. 3) and can create multiple
connections (HDBC).

The R3-Alpha codebase from which Ren-C's ODBC was adapted created
a new HENV for each database connection.  This changes to take
advantage of Ren-C's extension SHUTDOWN* callback to do a lazy
initialization of a global variable for henv which is cleaned up
on shutdown.

This removes the need for the HENV field in the object returned by
OPEN-CONNECTION.  At the same time, this also makes OPEN-CONNECTION
directly return an object instead of having Rebol code pass in a
pre-created object.  (The previous behavior was presumably because
it was difficult in R3-Alpha's C API to create objects in C from
scratch.)