Safe Shutdown ODBC if connection not closed
The ODBC extension is trying to bridge C's manual resource management
to a GC-oriented language. But it is possible for the user to write code
that opens a database connection that they don't close before shutdown.
This was triggering a panic() when the final shutdown GC started trying
to do ODBC cleanup after the ODBC extension (and the API services) had
all been unloaded.

Languages like Python have things like the WITH statement to help try
and keep users from ending up in this situation.  But mechanically it
can still happen so long as users are permitted to make manual calls
to open and close and omit the close, then shut down the system.

As a general category of problem there are likely more instances to be
looked at than just ODBC.  But it was creating problems, so this makes
use of the extension shutdown hook to iterate a separately tracked list
of database connections.  ODBC does not provide this list itself, so a
mirror list is kept.

More thought is needed, but this is just to try and avoid a panic() that
was interfering with a user's demo.