Handle Empty Windows GET-ENV variables
A mysterious error on a GitHub workflow pointed out that the Windows
API GetEnvironmentVariable() is extremely weird regarding an edge
case of variables that exist, but are not set to anything.

Usually if you say `SET VAR=` in a console this will unset a variable
but it is possible to programmatically create a variable that is set
yet has zero length.  Such variables are apparently made by GitHub
Actions workflows in the `env:` section.

Obscuring this is a wonky design of GetEnvironmentVariable(), which
takes an input buffer and a size.  If the buffer is insufficient in
size then it returns the necessary length including null terminator,
reserving 0 for errors.  But if the buffer is of sufficient size, it
returns the necessary length without a null terminator, hence 0 can
be a correct answer in that case.  (Why...would...they...do...that?!)

This adjusts the code to this annoying reality.  At the same time,
the opportunity is taken to make the environment var routines more
API reliant vs. using core types like REBCTX*