Purge 0.3.01 forms for 0.3.1
The original systems.r in Rebol would use leading zeros in the
third position for single digit variants.  This caused problems
when the tuple was rendered as all TUPLE! will be canonized to
not have leading zeros, getting it out of sync with things like bash
scripts that are just treating it like a string.