Multi-Return Improvements (Enfix, Meta)
Previously a multi-return could not be used with an enfix operation with
no GROUP! around the multi-return, which meant the overall expression
result of an enfix chain could not be stored in the primary result
variable.  That *should* be able to work, so this accomplishes it via a
somewhat ugly means of passing a flag to the frame filling code to say
that it's okay to leave deferred enfix operations in place, because it
promises to run the feed code later.  Then it re-enters the evaluator
after calling the frame at the lookahead step.  Convoluted though this
may be, it appears to work well enough for now.

This also uses an alternative solution for getting the result of a multi
return as "meta".  A previous hack would look for a caret on the right
of the SET-BLOCK! and make a note that wanted a meta result.  But
this uses a more generalized result of marking the return variable
slots as META-WORD! (or plain caret), which means the immediate
right of the SET-BLOCK! can always the function that has the multiple
return values.