Allow use of @xxx in DO
Since TAG! is a datatype that can hold full string information, it is
favored for the new semantics of "relative to currently executing
script" instead of the FILE! semantics of "relative to current dir".

This means `import @json` is being considered as the alternative
to `do <json>`.  Support it with DO as well.