Move ADJUST-URL-FOR-RAW from web repl to core
The functionality for being able to DO a URL of a GitHub or GitLab page
for a file and have it automatically redirect to a raw link was conceived to
make the web build more user friendly.  However, this behavior is just generally useful for the core
executable as well.

This migrates that URL translation into SYS.