PARSE2: New Bootstrap Compatibility Strategy
As UPARSE takes shape, the goal was to slowly evolve native PARSE toward
the new and improved default conventions (e.g. SET-WORD! for assignment
instead of marking series positions).  Old functionality like marking
series positions could then be done e.g. with `pos: here`.

Some clunky ideas about making the word HERE a no-op rule in bootstrap
were used, so that PARSE could work in the bootstrap executable as well
as under these rules.  This doesn't make a lot of sense, given that
UPARSE plans to give Redbol compatibility with UPARSE2.

Hence this introduces a new mode for native PARSE, e.g. PARSE/REDBOL
which is specialized as PARSE2.  Then, the bootstrap executable calls
its LIB.PARSE UPARSE2 and things are managed that way.

So this puts in backwards compatibility functionality for PARSE2.  In
In addition to backpedaling the old bootstrap idea, this goes ahead
and adds `<here>` and `<end>` behaviors to PARSE, because it was
noticed that the JSON encoder uses the word "here" for a variable.
This seems reasonable (just as using END for a variable seems to be
reasonable).

Code is updated to the new conventions.  One side-effect is that
variables holding TAG! cannot match against block inputs unless
they are quoted.