Restore PARSE/UPARSE of URL/PATH/TUPLE
Because of their immutability and lack of a position, types like
URL! and PATH! and TUPLE! were stopped from being supported by PARSE.
Mechanically a position is necessary in the state, and also if they were allowed to
be edited with things like CHANGE they may become invalid (or even
invisible since they have no outer delimiters for empty content).

This makes PARSE and UPARSE implicitly alias them as BLOCK!s or TEXT!,
as the read-only underlying data.  If mutations are desired, then the
user has to make a copy and then make sure the modified version can
still form a valid URL! or ANY-SEQUENCE!.