Enable running URL scripts from CLI
This makes it so if you pass something that looks like a URL on the CLI
then it will run that script from the network.

It leaves some open questions about cases where you have a file on disk
that looks like a URL with a colon in the name
of a directory.

https://forum.rebol.info/t/1764