Resurrect Build Extensions as .DLL/.so
This delves into the beast that is rebmake to try and grok how it ever
built DLL or .so extensions.  (The feature existed, but was never done
on the Travis CI at the time.  So I've only seen code for it, never seen
it actually working.)

In order to make things a bit clearer this avoids splitting extensions
into a lot of different lists, and instead just puts a `mode` flag on
each extension to say if it's builtin or dynamic or unused.

Then instead of having two processing code loops for builtin and dynamic
extensions, there's one loop for all extensions...which just has a bit
different code for each.  The shared code is thus more obvious.

As usual, comments are added to help decipher what the original
intent seems to have been.