Simplify PICK/POKE Implementation Concept
R3-Alpha had a technique in GOB!, which was copied in the FFI, that let
path dispatch "see ahead" more than one step at a time in the path
evaluation.  So `gob.size.x` would let the GOB! consume both SIZE and X, even
though `gob.size` would yield a PAIR! cell.  This let the GOB! compressed the size to be
smaller than a PAIR! in storage.

The way this was done was by framing path dispatch recursively.  While
most datatypes would let the path processor work on each step in a chain
the types like GOB! would manually invoke a recursion of path processing
form within their implementation.  This was very much a thorn.

An attempt to formalize this process by means of passing an array of
"steps" through a chain of PICKs and POKEs proved to be very annoying
and very slow.  This commit rips that out and moves toward a concept
that embraces "cells" for what they are and makes PICK and POKE cell-centric
operations,
which are used to build the higher level SET and GET operations.

A hack is included to allow the "last mile" of a SET to run a different
function than POKE, which is used to keep PROTECT working, as it has so
far depended on the PD_XXX path mechanics as well as the puzzling
Resolve_Path() function that wasn't used anywhere else.