Don't use DEFAULT with SET-PATH!
The new rules for DEFAULT don't allow SET-PATH! on the left.  The
bootstrap shim could work it in, or it could be allowed for compatibility, but
Redbol's DEFAULT was not enfix to begin with.
So this just gets rid of the 2 superfluous uses with paths as
part of the bootstrap.