Restore PICK* hack for REDBOL-PATHS
There had been code facilitating an aspect of the REDBOL-PATHS
compatibility which was removed, but turns out to be needed.

This adds it back--but it is less efficient now due to specializing the
function for each refinement used.  (The original had the full list of
pending "steps" being picked, now it moves only one pick step at a time
under the new rules.)