Use OPT SOME vs WHILE in PARSE
Historical PARSE had ANY as a keyword meaning "ANY NUMBER OF", which
runs against ANY meaning "ANY ONE OF" as it is used in the rest of the
language.  It was a looping construct, when there is no loop implied by the
word ANY elsewhere.

Even in PARSE, if you wrote:

    parse block [some any-value!]

You see ANY used in its general "ANY ONE OF" form.  Not only that, but ANY
does not imply "zero or some".  If I asked "do you have any bananas" and
you said yes, then it would be weird if I asked if I could have one and
you said "can't give you one, I don't have any".

This led Ren-C to standardize on WHILE by eliminating the difference
between ANY and WHILE, moving the progress requirement of ANY into a
separate keyword called FURTHER.  But the single arity WHILE was
confusing juxtaposed against the idea of a WHILE that has a body.

This transitions uses of WHILE to OPT SOME to see how that feels.

A few bootstrap usages cannot be converted because SOME has a progress
requirement that is no longer in Ren-C.