Don't use CTX() cast on inaccessible varlist
When advanced checking is turned on, casts like CtX() will ensure
that the pointer given is a valid context.  In the case of when a RETURN
is processing its binding, sometimes a varlist for a frame will be in
an inaccessible state.  These instrumented builds will panic if the cast is
attempted on an inaccessible series.

To get the desired result (a failed comparsion to the inaccessible
series) the thing being compared with should have its varlist extracted,
not coerce the stale varlist to a context.