Enable address sanitizer MSVC on Windows
Since the initial introduction of Address Sanitizer in MSVC in 2019, it
seems that it's gotten better as of March 2021.  It works on 64 bit
builds now, and works with debug builds, and doesn't require you to
do odd linking to i386 clang libraries...you just turn it on and it does
appear to work.

Given this, the hacky d-winstack.c mechanism that had been used in
a pinch is no longer necessary.  This incorporates the switches.