Fix ZIP creation of new files, better errors
In the libuv migration when the opening flags were supplied they were
mapped precisely so that `open/write` does not convey `/new`.  The ZIP
code was not using /NEW so it would not create new files.  This adds
the refinement while the question of what it should do is decided.

Also adds better errors when you write something like:

    >> zip %something.zip ["foo"]

This does not have a name for the file that will be extracted by UNZIP,
so you would expect something more like:

    >> zip %something.zip [%foo.txt "foo"]

The error message for this case is now better.