Hack for supporting dots in email addresses
It has always been the case that the scanner needed a full rewrite to
work as more of a clear state machine instead of ad-hoc code.  This has
only become more true as Ren-C has changed the concept of what paths
and tuples are, and added various new types and forms.

A tricky issue arose with how generic tuple processing was hacked into
the scanner to match how paths were scanned.  This led to email address
processing where if there were dots in the email address to the left of
the @ it would not be able to be loaded, since it would think of that
as a TUPLE! with an email address in it.

However as rewriting the scanner is not a project priority in exploring
the design, this hacks in an exception to be able to load in just that
case.  Rather than failing a tuple creation, it delimits it with a
dot and turns it into an email in this case.