Make ~void~ evaluate to pure invisible state
This switches to where a ~void~ BAD-WORD! specifically vaporizes (as
if it were the same behavior as (COMMENT "HI").  This eliminates the
need for so-called "impure invisibility" in functions like ALL and
ANY and REDUCE to skip over, because in the limit of this approach
there are no ~void~ isotopes.  However, the ~void~ BAD-WORD! can be
retained as the "meta" conveyance of this notion of pure invisibility.