Get rid of ~stale~ flag concept
The concept of impure invisibility put a burden on all constructs
that wanted to honor it.  It is likely the case that ~none~ can be
promoted to a kind of impure invisibility, with a construct like
DEVOID that turns them into pure invisibles for a similar effect
when it is needed...but the evaluator does not need to be involved.