Propagate ~none~ isotope as proxy for ~void~ isotope
The decision to make ~void~ the true "meta" representation of pure
invisibility means that there are no actual ~void~ isotopes, as their
evaluation vaporizes.  This means the previous idea of "conceptual
invisibility" being represented by a ~void~ isotope is now replaced
with "conceptual nothingness" as represented by ~none~ isotopes.

This means the previous idea that ^(comment "hi") would give back
a ~void~ isotope instead of a plain ~void~ (the way ^(~void~) would
have) is gone.  Now ^(comment "hi") gives back a plain ~void~, which
can be UNMETA'd back to invisibility.

With the closing of this representational gap, the ~none~ isotope takes
on the responsibility of being the result of things like `do []`,
`all []`, and `while [false] [<ignored>]`.  This propagates that around
the system as the new rule.