Redefine EVALUATE to swap its multi-return results
An initial conception of EVALUATE was that it would be common for
people to want to know if the block had reached the end, so the
main return result was made the block with the result as the secondary
value.  The idea got stretched to use quoting to annotate the block
return result when the evaluation had produced a value that was
invisible.

3 years later quoting was used for invisibility indication, but in a
more generic way through the ^META protocol.  And multi-returns became
flexible enough to allow you to mark any result as the expression's
overall result with a single @ character.

So this reacts by making EVALUATE return the updated block position as
the second result.  It also gives it a /VOID refinement to indicate
that vanishing is an acceptable result, so that individual evaluation
steps that vanish can be processed...otherwise they will be
automatically glossed over.