Make BAD-WORD!s truthy
Prior to the invention of isotopes, BAD-WORD!s were conceived as an
"ornery" datatype.  They still are somewhat ornery in the sense that
various places won't accept them (you can't APPEND a BAD-WORD! directly
to a block, you have to quote it, for instance).  But making them
truthy offers particular advantages by virtue of how then the only
result you can get from a ^META operation that is falsey is NULL.

This allows one to write simple solutions to classically difficult
problems, such as a FOR loop that iterates over two blocks and will
BREAK out of the loop correctly--and return the correct loop result:

    for-both: func ['var blk1 blk2 body] [
        unmeta all [
            meta for-each :var blk1 body
            meta for-each :var blk2 body
        ]
    ]

Only a NULL result from the FOR-EACH will remain NULL after the META
operation.  The other results will either be a plain BAD-WORD! or a
QUOTED!, both truthy, and both can be UNMETA'd back to their values.

This extremely compelling case outweighs any perceived value to making
the BAD-WORD! type either falsey or trigger an error when tested for
logical truth or falseness.  The "ornery" behavior is now part of
the isotopic form of BAD-WORD!.