Return ~quit~ isotope from plain QUIT, throw/catch isotopes
If a script calls QUIT and passes no result, then that is a normal
exit if it bubbles up to the console.  But if a script tries to use
the evaluative result of another script that just said QUIT (passed
no argument) then it is nice to know what happened.

Plain QUIT thus returns a ~quit~ isotope to help understand what
happened.  Also, THROW and CATCH are modified to be able to throw
isotopes since the mechanics have been added.