ENTRAP uses ^META quoting instead of values in blocks
When ENTRAP was invented, there were no isotopes...so it was possible
to put any non-null values into a block.  An error in a block was
thus distinguished as being a returned value, and an error not in a
block was a raised error.  This was particularly useful in the API
where it's currently difficult to get back more than one value as
a return result.

With isotopes and ^META conventions, it makes more sense to union the
potentially-null raised error with a ^META of the result value.  This
is not only less expensive than a BLOCK!, but it covers isotopes as
well.

The new ENTRAP function is usermode, and not a separate implementation
from TRAP.