Replace JS usage of reb.Void() with reb.None()
The concept of how to implement invisibility signals in the API has
not been worked out, as it does not correspond to a value type.

The API reb.Void() was being used in the JS bridge in order to
represent the "undefined" type.  This could be changed to actually
use an ~undefined~ isotope to be explicit about the JavaScript intent
but for now it just uses reb.None() instead.