Get rid of "Endish Nulled"
Prior to ^META parameters, there was no way to have an evaluative
`<end>` that could be distinguished from an evaluative NULL, because
both were represented by NULL.  With ^META, the unique state of ~void~
can only happen at the end of an input.

This means a misshapen idea called "endish nulleds" can be eliminated.
That concept was that the evaluated-or-unevaluated bit would be used,
where a "non-evaluated NULL" could be type checked as an END.  This
was a buggy bad idea, and it's good that it's gone.