Address some UPARSE and invisibility issues
The new concept that ~void~ isotopes vaporize instantly and are not
a distinct state from true invisibility shook the system around
somewhat.  UPARSE code actually fared pretty well, but there was one
piece where pure invisibility was not in the signature of a wrapper
function...so operations like `[remove skip]` were not working.

This addresses that issue, but also makes it clearer that ~none~
isotopes are the default products of rules that are not explicitly
proxies for invisibles.