Fix broken log() facility in make.sh
It was not doing what it was supposed to...instead it was
creating a file named `3`, due to not being able to do the
substitution that was attempted.

It's basically not possible to do what the code was trying:

https://stackoverflow.com/a/35556881

The attempt to document the somewhat cryptic 1>&3 is a drop
in the bucket compared to everything else, so this changes
to just make a practice of leaving a comment to see %log.sh
if someone does not know what that means.