New define for DID as isotope-friendly NOT NULL?
DID was introduced as a way of producing a LOGIC! from functions
that returned NULL in order to indicate "soft failure".  This
was to help avoid creating LOGIC!-bearing variations of the
functions like FIND?.

Its initial definition was as a synonym of TO-LOGIC, but it makes
more sense as a test for non-nullness.  This way something like
DID SELECT can be more meaningful.

With the advent of UPARSE, the parse function is now geared up to
return arbitrary results...and these may be isotopes (e.g. if the
parse returns `false`, it returns a ~false~ isotope to warn that
you can't use `if uparse` in that case and know that it didn't
complete the parse).  That makes DID ideal for being willing to
accept isotopes and return true...which solves a number of
historical problems.

In order to catch mistakes with the new semantics, DID will
give an error if it's passed a LOGIC! or BLANK!.  This is an
interim measure and will be taken out once the cases of DID
being used to mean TO-LOGIC have been changed.

It also introduces changes to the libRebol API where this is
concerned, introducing rebUnboxLogic() as a replacement for
most cases of rebDid()...but also bringing in rebTruthy() if
it is needed.