Give BLANK! a combinator, return NULL in UPARSE
In the idea of "everything is overrideable", the behavior of BLANK!
should be as well.  This takes the hardcoded handling of BLANK! and
turns it into a combinator that is a no-op that returns NULL.

It also enforces that UPARSE will turn all falsey return values into
their isotopes, which is likely a good general policy for most of
the functions that return `[<opt> any-value!]`.