Eliminate UPARSE?/PARSE?/PARSE2?, update tests
This is a very large change which embraces the new DID and DIDN'T
concept, as opposed to needing to make LOGIC!-bearing variations
of functions that return [<opt> any-value!] where you might want
to test the NULL case as soft-failure.

But intead of rotely converting the tests based on UPARSE? to just
check DID UPARSE?, this painstakingly codified what the actual
successful return result was.  It was done by hand as an audit of
any unusual results, which triggered some modifications.  While
painful to go through the 1,000 or so cases, it should provide
more test coverage.