Remove legacy MAYBE preparing for new meaning
Historical MAYBE was a DEFAULT-like construct which would evaluate
an expression on the right, and then update the variable on its left
if the expression was not null.  Otherwise it would GET the SET-WORD!
on the left and evaluate to the old value...leaving it alone.

While DEFAULT has to work with a left-enfix-quoting mechanism, a new
idea is that when functions on the right hand side of a SET-WORD!
evaluate to void, they do not update the value...and return the old one.

    >> x: 10
    == 10

    >> x: comment "so this would leave x alone"
    == 10

This allows constructs like MAYBE to be "opportunistically invisible",
and be useful when combined with a SET-WORD! or not.

However, the default MAYBE is planned on passing through NULL and
only being invisible on "invisible intent".  In light of this
semantic change