Coarse implementation of SET! of VOID as no-op
The initial conception of invisibles was that if they happened on
the right hand side of a SET-WORD!, there'd be a loop that would
keep going until it found a value:

    >> x: 1020

    >> x: comment "skipped" comment "me too" 1 + 2
    == 3

    >> x
    == 3

A number of technical questions and problems arose from this model,
which led to concern and head-scratching.  To demo just one of the
many problems:

    x: elide print "It's confusing when X gets [a]"
    append [] 'a

The idea of erroring was entertained, but ultimately a lightbulb
went on and the realization that a single-step evaluation on the
right hand side of a set could notice voidness and *just not
change the value*.

    >> x: 1020

    >> x: comment "skipped" comment "me too" 1 + 2
    == 3

    >> x
    == 1020

Without belaboring it too much in a commit message, this fits into
a larger model of how invisibility can be used as a kind of ternary
out-of-band state...added to the pure NULL mechanisms tested by
DID and DIDN'T.

This takes a very coarse first step at implementing the rule, which
requires an implementation change to LET.  This means LET taking a
SET-WORD! or SET-BLOCK! makes itself variadic and arity-2, and assumes
the responsibility for evaluating the expression (vs binding the
next value and then vaporizing itself).  It's par for the course in
terms of what kinds of hacks are being deemed acceptable in LET, so
all right to tweak in order to start bringing the clean rule into
existence.