Input /DECAY option for THEN/ELSE/ALSO, no @[...] branches
Prior to the existence of ~null~ isotopes, a proposal to subvert the
"voidification" of branches was made using @[...].

    >> if true [null]
    == #[void]

    >> if true @[null]
    ; null

The advent of ~null~ isotopes means that in general, nulls can be
tunneled out of branch constructs.  And holistically speaking, it is
better to guarantee that pure NULL is reserved completely for the
case of no branch running in a control construct.  This permits the
building of higher-level control constructs out of lower-level ones,
much as reserving pure NULL for BREAK does in looping constructs.

This removes the @[...] feature for branches, and instead adds new
variants for functions like ELSE and THEN which will react to the
isotopic ~null~ in the same way as if it were not isotopic.  This
moves the intention of reacting to the branch result on the left
into the hands of the construct consuming the branch product--not
on the producer--which is much better.

    >> if true [print "if", null] else [print "else"]
    if
    ; isotopification of the null prevents ELSE from running

    >> if true [print "if", null] *else [print "*else"]
    if
    *else
    ; the *else construct considers ~null~ isotopes to be null