Make THE handle BAD-WORD!, unlike @
For ease-of-use in the API, the function assigned to @ was made to
disallow all BAD-WORD! arguments except for ~null~.  This reserved
that for being able to make pure NULL.

    >> @ (1 + 2)
    == (1 + 2)

    >> @ word
    == word

    >> @ ~something~
    ** Error: @ only accepts the BAD-WORD! of ~null~ to make NULL

    >> @ ~null~
    ; null

The motivation is that if you are splicing values into a block of code,
and you have a NULL, it cannot be reified into the block.  It has to be quoted.  But as a workaround, the API splices NULL as a ~null~ BAD-WORD!, and you can operate on it with @:

    REBVAL *var = nullptr;
    if (rebUnboxLogic("null?", rebQ(var)) { ... }
    if (rebUnboxLogic("null? @", var) { ... }  // cleaner!

So long as you don't genuinely have a BAD-WORD! in var, this works out.
The error cues you hopefully to avoid errors in most cases (unless you
really incidentally only have a ~null~ BAD-WORD! and want it literally).

@ was synonymous with the word THE.  However, THE is useful with
BAD-WORD! arguments for certain readability scenarios...and there's
no real reason why @ has to be exactly the same.

This makes THE work on BAD-WORD!s and removes the API-friendly quirk:

    >> the ~something~
    == ~something~

    >> the ~null~
    == ~null~