Implement "blackhole in, isotope out" for APPEND, etc.
This introduces a new `<blackhole>` tag for use in parameter specs,
that will let a function take a `#` in and produce a ~blackhole~
isotope as an output.

The new behavior is added to APPEND so that there is a way of
opting out of the series to append to, which is not too easily
conflated with the default state of empty variables.