Deprecate arity-1 WHILE in UPARSE/PARSE3
The word ANY was deemed a poor choice for what it was doing (matching
"any" number of times--zero or more) because the general usage of the
word ANY in Rebol has been to pick one item out of a set of them.
(ANY-VALUE!, or the ANY function).  It did not suggest looping.

So it was tried to standardize on the arity-1 WHILE construct of
Rebol3 and Red instead...breaking out the progress-requirement nuance
into its own special primitive of FURTHER.

This resulted in the unsettling appearance of an arity-1 construct,
so it was tried to make WHILE in the language arity-1 to match UNTIL,
so that parse code would seem more consistent with everything else.
The arity-2 form was given to LOOP.

Ultimately this was deemed unsatisfactory, and the benefits of breaking
out a term to indicate optionality (MAYBE or OPT) offered a new and
interesting nuance consistent with the rest of the system.  Also, the
possibility of an arity-2 WHILE in PARSE shows a construct with high
potential and much less confusing.

Additionally--it was found that many people were using ANY/WHILE for
"zero or more matches" in places where one-or-more-matches would always
be present.  This "laziness" cost nothing and made the program less
clear.  Breaking out the optionality into its own component helps
see the real structure behind the rules, and allows interesting
refactoring tied to the choice of OPT vs. MAYBE (or other tools).

So this deprecates WHILE in both UPARSE and PARSE3.  The compatibility
parse used by bootstrap (PARSE2) still offers it.