No non-interstitial invisibles in non SET-WORD! cases
"Interstitial invisibles" are those that aren't on evaluative
step boundaries.  The idea of allowing these was in order to
permit commenting out arbitrary ranges of code, e.g.

    >> 1 + comment [2 *] 3
    == 4

While an interesting idea, this has not turned out to be the main
thrust of what invisibles are used for.  The convolutions required
to achieve a semi-working version of the above were brittle and
not applicable to dialects which interlock parts in a predetermined
structure (such as the linking of combinators to make parsers in
UPARSE).

This removes the ad-hoc mechanisms for iterating during an argument
evaluation to try and accommodate retriggering in the case of an
invisible that is missing.