Change ~ to be unset isotope, no trash in release build
After having some experience with it, ~ is the better representation
for an unset isotope.

* Reduces noise in FRAME! to see which variables specialized

* Less chance for confusion since UNSET? takes a variable;
  if it were named ~unset~ people would likely expect
  `(unset? ~unset~)` to work.

* Quick way to unset variables, simply `(var: ~)`

So this makes it so a nullptr label inside a BAD-WORD! represents the
unique nameless state.

That state had previously been used for indicating GC-unsafe "Trash",
but that is now done with NODE_FLAG_FREE on the BAD-WORD!.  This also
changes it so that the release build just puts the unset isotope with
no flag in places where trash is initialized, which means IS_TRASH()
can be false and not cost anything in the GC.