Implement GET-GROUP! as a combinator
Initially GET-GROUP! was a feature of the BLOCK! combinator, which
meant it could have variadic behavior, with the spliced rule
material picking up an argument:

    >> uparse "aaa" [:(second [not some]) "a"]
    == "a"

However, that meant GET-GROUP! did not dispatch to a combinator,
making it not possible to override the behavior (or to borrow the
behavior for another ANY-ARRAY! type).

This was noticed while trying to show UPARSE being reconfigured to
act like an "ugly" experiment that was tried.  The experiment had
plain GROUP! would splice its result the way a GET-GROUP! currently
does...and you would have to ELIDE it to get it not to.  The demo
transfers the GET-GROUP! behavior to GROUP!.

This includes that demo with the change.