Unify "none" and "unset" isotopes
The ~ isotope is now used for unset variables, and there are some
plausibly good reasons for making it the isotope that the console
does not show any output for.  It also needs a name, and calling
it "none" is good and short...which meshes with the fact that
no output is printed for it.

This leads to some common behavior with historical Rebol where
the thing that isn't printed in console output is the same thing
that is held by variables that are not set.