Fix stack movement issue in librebol test code
Building with clang and running the tests revealed that there was
a stack pointer being held across an API call.  This code pattern
is not legal:

    Init_Integer(DS_PUSH(), rebUnboxInteger(...));

The problem is that Init_Integer() is receiving a direct pointer into
the data stack, and the API call does pushes and pops to the stack.
There's no guarantee on parameter ordering in the compiler--so this
will be a problem if the left gets evaluated first...as the pushed
address has to be held across the API call.